requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.1
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.5.0
